fix: handler add error when directory not exist

+++ src/basic-operations/add.js
@@ -2,13 +2,18 @@ import fs from "fs";
 import path from "path";
 import { ConsoleOutput } from "../helpers/constants.js";
 
-const { fileExist } = ConsoleOutput;
+const { fileExist, noDirectory } = ConsoleOutput;
 
 export const add = async (currentDirectory, toRead) => {
   const newPath = path.resolve(currentDirectory, toRead);
 
-  fs.writeFile(newPath, "", { flag: "wx" }, (err) => {
-    if (err) console.log(fileExist);
-  });
+  if (fs.existsSync(newPath)) {
+    fs.writeFile(newPath, "", { flag: "wx" }, (err) => {
+      if (err) console.log(fileExist);
+    });
+  } else {
+    console.log(noDirectory);
+  }
+
   console.log(`You are currently in ${currentDirectory}`);
 };
